move useUserTenantPermissions to tmp folder
