+++ src/tmp/hooks/index.js
@@ -1,0 +1,1 @@
+export { default as useUserTenantPermissions } from './useUserTenantPermissions';
